Do not delete TeamDrive if missing from Perun by default
- When TeamDrive is missing in input file from Perun, just remove all
  permissions and do not delete the Drive.
- Added new config option "allow_delete_teamdrive". If set to "true",
  TeamDrive is deleted.

+++ src/main/java/cz/metacentrum/perun/googlegroupconnector/GoogleGroupsServiceImpl.java
@@ -936,9 +936,14 @@ public class GoogleGroupsServiceImpl implements GoogleGroupsService {
 
 				if (perunDrive == null) {
 
-					// delete domain drive (not found from Perun)
-					deleteTeamDrive(domainDrive);
-					teamDrivesDeleted++;
+					if (Boolean.getBoolean(properties.getProperty("allow_delete_teamdrive", "false"))) {
+						// delete domain drive (not found from Perun)
+						deleteTeamDrive(domainDrive);
+						teamDrivesDeleted++;
+					} else {
+						// keep teamdrive but remove all permissions
+						processTeamDrivePermissions(domainDrive, new ArrayList<>());
+					}
 
 				}
 			}
